my-external-api-project/requirements.txt
Flask==3.0.3 # Or current stable version
psycopg2-binary==2.9.9 # Or current stable version
python-dateutil==2.9.0 # If you end up using it for date parsing/formatting